backend/requirements.txt
# Django + DRF
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-notifications-hq==1.8.2

# Authentification
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0  # JWT pour dj-rest-auth
dj-rest-auth[with_jwt]==5.0.2


# CORS
django-cors-headers==4.3.1

# Database & configuration
# psycopg2-binary==2.9.7
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0  # pratique pour DATABASE_URL

# Images
Pillow==11.0.0

# Storage (S3-compatible: Cloudflare R2)
django-storages[s3]==1.14.2
boto3==1.34.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0  # pour brancher Redis comme cache Django

# WebSockets / Channels
channels==4.1.0
channels-redis==4.2.0

# ASGI server (pour servir HTTP + WebSockets)
daphne==4.1.0

# Dev / utilitaires (optionnel)
whitenoise==6.5.0  # servir les statiques en prod sans nginx
drf-spectacular==0.27.2  # docs OpenAPI/Swagger
gunicorn==23.0.0


# Logging
sentry-sdk[django]==2.20.0

# Runtime tooling (provides pkg_resources used by some libs)
setuptools>=75.0.0